+++ Graph_Architecture/Vertex.h
@@ -1,7 +1,3 @@
-//
-// Created by luca eaton on 3/11/26.
-//
-
 #ifndef VECTOR_H
 #define VECTOR_H
 #include <vector>
@@ -10,7 +6,7 @@ using namespace std;
 
 class Vertex {
 private:
-    size_t id_;
+    long long id_;
     double lat_;
     double lon_;
     vector<Edge*> edges_;
@@ -19,7 +15,7 @@ public:
     Vertex(long long i, double la, double lo);
     void addEdge(Edge *edge);
     [[nodiscard]] const std::vector<Edge *> &getEdges() const;
-    [[nodiscard]] size_t getId() const;
+    [[nodiscard]] long long getId() const;
     [[nodiscard]] double getLat() const;
     [[nodiscard]] double getLon() const;
 };
